Added injectable in controller made translate an exported function
- Simon D
